+++ src/interactivityservice.ts
@@ -116,9 +116,6 @@ module powerbi.extensibility.utils.interactivity {
         /** Checks whether the selection mode is inverted or normal */
         isSelectionModeInverted(): boolean;
 
-        /** Apply new selections to change internal state of interactivity service from filter */
-        applySelectionFromFilter(appliedFilter: filter.AppliedFilter): void;
-
         /** Apply new selections to change internal state of interactivity service */
         restoreSelection(selectionIds: ISelectionId[]): void;
     }
@@ -140,6 +137,39 @@ module powerbi.extensibility.utils.interactivity {
         applySelectionFilter(): void;
     }
 
+    export interface IExtensibilityMeasuredSelecionId extends ExtensibilityISelectionId {
+        dataMap: SelectorsForColumn;
+        measures: string[];
+    }
+
+    export interface IMeasuredSelectionId extends ISelectionId {
+        dataMap: SelectorsForColumn;
+        measures: string[];
+        compareMetadata(currentDataMap: SelectorsForColumn, otherDataMap: SelectorsForColumn): boolean;
+        compareMeasures(currentMeasures: string[], otherMeasures: string[]): boolean;
+    }
+
+    export interface SelectorsForColumn {
+        [queryName: string]: data.DataRepetitionSelector[];
+    }
+
+    // It's a temporary function for compatibility with API 2.1
+    // It will probably be removed after API 2.2 release
+    export function checkDatapointAgainstSelectedIds(dataPoint: SelectableDataPoint, selectedIds: ISelectionId[]) {
+        return selectedIds.some((value) => {
+            const measuredValue: IMeasuredSelectionId = value as IMeasuredSelectionId;
+            const otherSelectionId: IExtensibilityMeasuredSelecionId = dataPoint.identity as IExtensibilityMeasuredSelecionId;
+            if (measuredValue.dataMap && otherSelectionId.dataMap && measuredValue.compareMetadata(measuredValue.dataMap, otherSelectionId.dataMap)) {
+                return true;
+            }
+            if (!measuredValue.dataMap && measuredValue.compareMeasures(measuredValue.measures, otherSelectionId.measures)) {
+                return true;
+            }
+
+            return false;
+        });
+    }
+
     export class InteractivityService implements IInteractivityService, ISelectionHandler {
         private selectionManager: ISelectionManager;
 
@@ -151,8 +181,6 @@ module powerbi.extensibility.utils.interactivity {
         // Selection state
         private selectedIds: ISelectionId[] = [];
         private isInvertedSelectionMode: boolean = false;
-        private hasSelectionOverride: boolean;
-        private behavior: any;
 
         public selectableDataPoints: SelectableDataPoint[];
         public selectableLegendDataPoints: SelectableDataPoint[];
@@ -193,26 +221,18 @@ module powerbi.extensibility.utils.interactivity {
                     this.selectableDataPoints = dataPoints;
                     this.renderSelectionInVisual = () => behavior.renderSelection(this.hasSelection());
                 }
-
-                if (options.hasSelectionOverride != null) {
-                    this.hasSelectionOverride = options.hasSelectionOverride;
-                }
-
             }
             else {
                 this.selectableDataPoints = dataPoints;
                 this.renderSelectionInVisual = () => behavior.renderSelection(this.hasSelection());
             }
 
-            // Bind to the behavior
-            this.behavior = behavior;
             behavior.bindEvents(behaviorOptions, this);
             // Sync data points with current selection state
             this.syncSelectionState();
         }
 
         private clearSelectedIds(): void {
-            this.hasSelectionOverride = undefined;
             ArrayExtensions.clear(this.selectedIds);
         }
 
@@ -234,17 +254,10 @@ module powerbi.extensibility.utils.interactivity {
             }
 
             for (let dataPoint of dataPoints) {
-                dataPoint.selected = InteractivityService.isDataPointSelected(dataPoint, this.selectedIds);
+                dataPoint.selected = checkDatapointAgainstSelectedIds(dataPoint, this.selectedIds);
             }
 
             return this.hasSelection();
-        }
-
-        /**
-         * Apply new selections to change internal state of interactivity service from filter
-         */
-        public applySelectionFromFilter(appliedFilter: filter.AppliedFilter): void {
-            this.restoreSelection(filter.FilterManager.restoreSelectionIds(appliedFilter));
         }
 
         /**
@@ -381,7 +394,7 @@ module powerbi.extensibility.utils.interactivity {
             }
 
             selectableDataPoints.forEach((dataPoint: SelectableDataPoint) => {
-                const shouldDataPointBeSelected: boolean = !InteractivityService.isDataPointSelected(dataPoint, originalSelectedIds);
+                const shouldDataPointBeSelected: boolean = !checkDatapointAgainstSelectedIds(dataPoint, originalSelectedIds);
 
                 this.selectSingleDataPoint(dataPoint, shouldDataPointBeSelected);
             });
@@ -474,17 +487,13 @@ module powerbi.extensibility.utils.interactivity {
             let foundMatchingId = false;
 
             for (let dataPoint of selectableDataPoints) {
-                dataPoint.selected = InteractivityService.isDataPointSelected(dataPoint, selectedIds);
+                dataPoint.selected = checkDatapointAgainstSelectedIds(dataPoint, selectedIds);
 
                 if (dataPoint.selected)
                     foundMatchingId = true;
             }
 
             return foundMatchingId;
-        }
-
-        private static isDataPointSelected(dataPoint: SelectableDataPoint, selectedIds: ISelectionId[]): boolean {
-            return selectedIds.some((value: ISelectionId) => value.includes(dataPoint.identity as ISelectionId));
         }
 
         private removeSelectionIdsWithOnlyMeasures() {