+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.0.0
+* REM: function restoreSelectionIds from filter manager because it's unsupportable anymore with API 2.1
+* REM: function applySelectionFromFilter from service because it's unsupportable anymore with API 2.1
+* REM: function isDataPointSelected from service because it's incompatible with API 2.1 for a while
+* ADD: exportable function checkDatapointAgainstSelectedIds to provide compatibility with API 2.1
+
 ## 3.2.0
 * UPD: ability to select multiple data points via `handleSelection` method